Fix JRuby tests where we can't Kernel#fork. Re. c7c82108faba872fbecd00320a0f1b04e3140e0a

+++ test/helper.rb
@@ -2,22 +2,28 @@ require 'minitest/autorun'
 
 module Rack
   class TestCase
-    # Keep this first.
-    PID = fork {
-      ENV['RACK_ENV'] = 'deployment'
-      ENV['RUBYLIB'] = [
-        ::File.expand_path('../../lib', __FILE__),
-        ENV['RUBYLIB'],
-      ].compact.join(':')
+    if `which lighttpd` && !$?.success?
+      begin
+        # Keep this first.
+        PID = fork {
+          ENV['RACK_ENV'] = 'deployment'
+          ENV['RUBYLIB'] = [
+            ::File.expand_path('../../lib', __FILE__),
+            ENV['RUBYLIB'],
+          ].compact.join(':')
 
-      Dir.chdir(::File.expand_path("../cgi", __FILE__)) do
-        exec "lighttpd -D -f lighttpd.conf"
+          Dir.chdir(::File.expand_path("../cgi", __FILE__)) do
+            exec "lighttpd -D -f lighttpd.conf"
+          end
+        }
+      rescue NotImplementedError
+        warn "Kernel#fork unsupported. Skipping CGI and FastCGI tests."
+      else
+        Minitest.after_run do
+          Process.kill 15, PID
+          Process.wait(PID)
+        end
       end
-    }
-
-    Minitest.after_run do
-      Process.kill 15, PID
-      Process.wait(PID)
     end
   end
 end
